improving stream/services initialization and servlets   (#16)
* updating haystack-commons version

* updating documentation

* adding config for ktable changelog topic, making changelog in-memory store and documenting configurations

* gracefully handling instantiation and startup exceptions and corresponding teardowns

* structuring resources to use common base class

+++ graph-builder/src/main/scala/com.expedia.www.haystack.service.graph.graph.builder/config/entities/ServiceClientConfiguration.scala
@@ -22,7 +22,7 @@ package com.expedia.www.haystack.service.graph.graph.builder.config.entities
   * @param connectionTimeout
   * @param socketTimeout
   */
-case class ServiceClientConfiguration(connectionTimeout: Long, socketTimeout: Long) {
+case class ServiceClientConfiguration(connectionTimeout: Int, socketTimeout: Int) {
   require(connectionTimeout > 0)
   require(socketTimeout > 0)
 }
